Refactor asset list logic to be located in a repository class.
