+++ app/src/main/java/tech/ula/viewmodel/FilesystemEditViewModel.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.database.sqlite.SQLiteConstraintException
 import kotlinx.coroutines.experimental.launch
-import tech.ula.model.AppDatabase
+import tech.ula.model.repositories.AppDatabase
 import tech.ula.model.entities.Filesystem
 import tech.ula.utils.async
 import kotlin.coroutines.experimental.Continuation
