+++ app/src/main/java/tech/ula/viewmodel/SessionEditViewModel.kt
@@ -5,7 +5,7 @@ import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.database.sqlite.SQLiteConstraintException
 import kotlinx.coroutines.experimental.launch
-import tech.ula.model.AppDatabase
+import tech.ula.model.repositories.AppDatabase
 import tech.ula.model.entities.Filesystem
 import tech.ula.model.entities.Session
 import tech.ula.utils.async
